+++ src/integrationTest/groovy/io/github/devatherock/emailsender/controllers/EmailControllerIntegrationSpec.groovy
@@ -1,0 +1,18 @@
+package io.github.devatherock.emailsender.controllers
+
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.test.context.ContextConfiguration
+
+import io.github.devatherock.emailsender.test.IntegrationTestApp
+
+/**
+ * Integration test for {@link EmailController} class
+ */
+@ActiveProfiles('integration')
+@SpringBootTest(classes = [IntegrationTestApp],
+        webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@ContextConfiguration
+class EmailControllerIntegrationSpec extends EmailControllerSpec {
+
+}