+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#6](https://github.com/devatherock/email-sender/issues/6): Unit tests
 - Dependency check plugin
 - Used [Wiser](https://github.com/voodoodyne/subethasmtp/blob/master/Wiser.md) for integration tests
+- [#8](https://github.com/devatherock/email-sender/issues/8): Integration tests
 
 ### Changed
 - Upgraded gradle to `7.6.1` and java to `17`